Fix per VM environment.
Fixes Java, as CLASSPATH was getting lost, and Graal as both CLASSPATH
and JAVA_CMD were getting lost.
